features for user avatar update.

+++ stage8/NotesApi/src/controllers/notesController.js
@@ -14,11 +14,11 @@
     class NotesController{
 
         async show(request, response){
-            const { id } = request.params;
+            const user_id = request.user.id;
 
-            const note = await knex("notes").where({ id }).first();
-            const tags = await knex("tags").where({ note_id: id}).orderBy("name");
-            const links = await knex("links").where({ note_id: id}).orderBy("created_at");
+            const note = await knex("notes").where({ id: user_id }).first();
+            const tags = await knex("tags").where({ note_id: user_id}).orderBy("name");
+            const links = await knex("links").where({ note_id: user_id}).orderBy("created_at");
 
             return response.json({...note, tags, links});
         }
@@ -26,7 +26,7 @@
        async create(request, response){
 
             const { title, description, links, tags } = request.body;
-            const { user_id } = request.params;
+            const { user_id } = request.user.id;
 
             const [note_id] = await knex("notes").insert({
                 title,
@@ -39,8 +39,6 @@
                     note_id,
                     url: link
                 }
-
-
             })
 
             await knex("links").insert(linksInsert);
@@ -60,16 +58,18 @@
        }
 
        async delete(request, response){
-            const { id } = request.params;
+            const user_id = request.user.id;
 
-            await knex("notes").where({id}).delete();
+            await knex("notes").where({id: user_id}).delete();
 
             return response.json();
        }
 
        async index(request,response){
 
-            const { user_id, title, tags} = request.query;
+            const { title, tags } = request.query;
+
+            const user_id = request.user.id;
 
             console.log(request.query);
